Move test to exercise higher level function
This is because the higher-level function takes into account OS-specific issues yet exercises
the same lower-level function eventually.
